'修改i静态化规则'

+++ controllers/BlogController.php
@@ -126,31 +126,37 @@ class BlogController extends BaseController{
         $arr = $model->getAllBlogHtml();
         mkdir(ROOT."views/html/blogs",777,true);
         foreach($arr as $v){
-            $id = $v['id'];
-            $data = $model->getBlog($id);
-            $add = $model->addLook($id);
-            if(!$data)
-                return false;
-            $pre =$this->getPre($id);
-            $next = $this->getNext($id);
-            $relevant = $model->getRele($data['cat_3']);
-            foreach($relevant as $k=>$v){
-                if($id == $relevant[$k]['id']){
-                    unset($relevant[$k]);
+            if(file_exists(ROOT."views/html/blogs/".$v['id'].".html")){
+                echo $v['id'].",";
+            }else
+            {
+                $id = $v['id'];
+                $data = $model->getBlog($id);
+                $add = $model->addLook($id);
+                if(!$data)
+                    return false;
+                $pre =$this->getPre($id);
+                $next = $this->getNext($id);
+                $relevant = $model->getRele($data['cat_3']);
+                foreach($relevant as $k=>$v){
+                    if($id == $relevant[$k]['id']){
+                        unset($relevant[$k]);
+                    }
                 }
+                $data = [
+                    "pre" => $pre?: null,
+                    "current" => $data?: null,
+                    "next" => $next?: null,
+                    "relevant"=>$relevant ? :null,
+                ];
+                ob_start();
+                view('info/info',[
+                    "data"=>$data
+                ]);
+                $str = ob_get_contents();
+                file_put_contents(ROOT."views/html/blogs/{$id}.html",$str);
             }
-            $data = [
-                "pre" => $pre?: null,
-                "current" => $data?: null,
-                "next" => $next?: null,
-                "relevant"=>$relevant ? :null,
-            ];
-            ob_start();
-            view('info/info',[
-                "data"=>$data
-            ]);
-            $str = ob_get_contents();
-            file_put_contents(ROOT."views/html/blogs/{$id}.html",$str);
+           
         }
         
     }
